refactor(py-examples): use DoIPMultiECUClient in all examples
Refactored all Python examples in `motion-server/examples/py/` to use the `DoIPMultiECUClient` instead of the `doipclient.DoIPClient`.

This change makes the examples consistent and ensures they all use DoIP protocol version 3, which is the default in `DoIPMultiECUClient`.

This also resolves an issue where some examples would fail due to using the default protocol version 2 of `doipclient`.